More theme provider refactoring. Now takes an initial theme and then provides a hook to change it after initialization. Separate theme list into its own package.
