+++ layout/index.js
@@ -1,3 +1,2 @@
 export * from './src/mdx-layout';
 export * from './src/story-layout';
-export * from './src/mdx-button-layout';
