+++ layout/src/story-layout.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
+import { themeList } from '~/theme/src/themeList';
 import { ThemeProvider } from '../../theme-provider';
+import { Themes } from '../../theme-list/themeList';
 
 export const StoryLayout = {
   decorators: [
     (Story) => (
-      <ThemeProvider>
+      <ThemeProvider initialTheme={themeList[Themes.Light]}>
         <div style={{ padding: '1rem' }}>
           <Story />
         </div>
