+++ theme-list/types/themeList.ts
@@ -1,0 +1,5 @@
+import { ConfigType } from '@stitches/react/types/config';
+
+export type ThemeList = {
+  [key: string]: ConfigType.Theme;
+};
