+++ theme-provider/index.ts
@@ -1,2 +1,1 @@
 export * from './context/ThemeProvider';
-export { ThemeMode } from './types/themeMode';
