--- theme-provider/types/themeListItem.ts
@@ -1,11 +1,0 @@
-import { ConfigType } from '@stitches/react/types/config';
-
-export type ThemeListItem = {
-  themeName: string;
-  theme: ConfigType.Theme;
-  className: string;
-};
-
-export type ThemeList = {
-  [key: string]: ThemeListItem;
-};
