--- theme-provider/types/themeMode.ts
@@ -1,4 +1,0 @@
-export enum ThemeMode {
-  Light = 'Light',
-  Dark = 'Dark',
-}
